Fixed up blending and started particle system
I fixed up a lot of junk with the blending. The GUI is perf now that the
depth test has been shut off, and the reset of the blending presumes
black as the default non-colour, as it should. Colours and alphas will
be added/used from there.

Started implementation of particle system. Not sure how I'm going to
incorporate lines, especially if motion is not perfectly linear. Make
them curves?